==> app/notebooks/notebook_01.ipynb <==
from datetime import datetime
import uuid
import pandas as pd
from sqlalchemy.orm import sessionmaker
from sqlalchemy import MetaData
import numpy as np
# from app.database.queries import insert_data
from typing import List, Dict
import logging

logger = logging.getLogger(__name__)  # Use a named logger

logging.basicConfig(level=logging.INFO)
RECIPE = "../../app/data/processed/recipe_cleaned.json"
INGREDIENTS = "../../app/data/processed/ingredients_AI.json"
MEAL = "../../app/data/processed/meal_cleaned.json"


logger.info("PREPARING DATA")
data: pd.DataFrame = pd.read_json(INGREDIENTS)
data = data.replace({np.nan: None, '': None, ' ': None})
data.head()

unique_ingredients = data[["main_ingredient"]].drop_duplicates()
unique_ingredients["ingredientId"] = range(12, len(unique_ingredients) + 12)
recipe_to_ingredients = data.merge(unique_ingredients, on="main_ingredient", how="left")
recipe_to_ingredients.head(3)

alternative_ingredient = data[["alternative_ingredient"]].drop_duplicates()
alternative_ingredient["alternativeOfId"] = range(1, len(alternative_ingredient) + 1)
recipe_to_ingredients = recipe_to_ingredients.merge(alternative_ingredient, on="alternative_ingredient", how="left")
recipe_to_ingredients.head(3)

# alt_ingredient = data[["alternative_ingredient"]].drop_duplicates()
# alt_ingredient["alternative_ingredient"] = alt_ingredient[alt_ingredient["alternative_ingredient"] != None]
# alt_ingredient["alternativeOfId"] = range(1, len(alt_ingredient)+1)
# recipe_to_ingredients = recipe_to_ingredients.merge(alt_ingredient, on="alternative_ingredient", how="left")
# recipe_to_ingredients.head(3)


measurements = data[["measurement"]].drop_duplicates().reset_index(drop=True)
# measurement_AI = pd.read_json("../../app/data/processed/measurements_AI.json")
measurements["measurementId"] = range(6, len(measurements) + 6)
recipe_to_ingredients = recipe_to_ingredients.merge(measurements, on="measurement", how="left")
len(recipe_to_ingredients)

recipe_to_ingredients = recipe_to_ingredients.rename(columns={"ingredient_id": "recipe_id"})
recipe_to_ingredients.drop(columns=['main_ingredient', "measurement", "weight", "alternative_ingredient"], inplace=True)
data_dict: List[Dict] = recipe_to_ingredients.to_dict(orient='records')  # type: ignore
print(data_dict)

import unicodedata
from fractions import Fraction

unicodes = recipe_to_ingredients[["quantity"]]
unicodes["id"] = range(1, len(unicodes)+1)

unicode_list = unicodes["quantity"]
unicode_id = unicodes['id']

unicode_json = []

def contains_unicode(input_str):
    return any(ord(char) > 127 for char in input_str)

for unicode, id in zip(unicode_list, unicode_id):
    full = 0
    description = ""
    unicode_dict = {}
    if unicode is not None:    
        if contains_unicode(unicode) or "/" in unicode:
            # print(unicode)
            unicode_str = unicodedata.normalize("NFKC", unicode).replace("⁄", "/")
            if ' ' in unicode_str:
                if "-" in unicode_str:
                    parts = unicode_str.split("-")
                    full = float(Fraction(parts[0].strip()))
                else:
                    parts = unicode_str.split()
                    whole_number = int(parts[0])
                    fraction_part = Fraction(parts[1])
                    full = whole_number + float(fraction_part)
            else:
                full = float(Fraction(unicode_str))
        else:
            try:
                full = float(unicode)
            except Exception as e:
                # print(e)
                if isinstance(unicode, str):
                    full = -1
                    description = str(unicode)
                    print(description)
        unicode_dict["id"] = id
        unicode_dict["value"] = full
        unicode_dict["description"] = description
    else:
        full = -1
        description = "None"
        unicode_dict["id"] = id
        unicode_dict["value"] = full
    
    unicode_json.append(unicode_dict.copy())
    
# print(unicode_json)

quantity_details = pd.DataFrame(unicode_json)
rti_copy = recipe_to_ingredients.copy()

rti_copy["quantity"] = quantity_details["value"]
rti_copy["quantityDescription"] = quantity_details["description"]
rti_copy

rti_copy = rti_copy.rename(columns={"recipe_id": "recipeId"})
rti_copy["id"] = range(22, len(rti_copy)+22)
rti_copy = rti_copy[["id","quantity", "measurementId", "ingredientId", "alternativeOfId", "recipeId", "quantityDescription"]]
rti_copy

data_dict = rti_copy.to_dict(orient="records")
# insert_data(data=data_dict, schema="RecipeIngredient")

# u_ingredients = unique_ingredients.rename(columns={"ingredientId" : "id", "main_ingredient": "name"})
# u_ingredients = u_ingredients[["id", "name"]]
# u_ingdict = u_ingredients.to_dict(orient="records")
# insert_data(data=u_ingdict, schema="Ingredient")

rti_copy

logger.info("PREPARING DATA")
data: pd.DataFrame = pd.read_json(RECIPE)
data = data.replace({np.nan: None, '': None, ' ': None})
data.head()
