==> app/notebooks/notebook_02.ipynb <==
from datetime import datetime
import uuid
import pandas as pd
from sqlalchemy.orm import sessionmaker
from sqlalchemy import MetaData
import numpy as np
from app.database.queries import insert_data
from typing import List, Dict
import logging
import logging
from typing import Dict, List
from sqlalchemy import MetaData
from sqlalchemy import text
from sqlalchemy.orm import sessionmaker
from app.database.connect import dbconnect

logger = logging.getLogger(__name__)  # Use a named logger

logging.basicConfig(level=logging.INFO)
RECIPE = "../../app/data/processed/recipe_cleaned.json"
INGREDIENTS = "../../app/data/processed/ingredients_AI.json"
MEAL = "../../app/data/processed/meal_cleaned.json"
MEALTORECIPE = "../../app/data/processed/mealtorecipe.json" 

logger.info("PREPARING DATA")
data: pd.DataFrame = pd.read_json(MEAL)
data = data.replace({np.nan: None, '': None, ' ': None})
data.head()

data["updated_at"] = data["created_at"]
data = data.rename(columns={"meal_id": "id"})

data["status"] = "PUBLISHED"
data["video_url"] = "Not Available"

data = data[["id", "name", "description", "status", "created_at", "updated_at", "source", "video_url"]]
len(data)

df_cleaned = data.drop_duplicates(subset='name', keep='first')
df_cleaned = df_cleaned.reset_index().drop(columns="index")

data = df_cleaned.copy()
data_dict = data.to_dict(orient='records')

# insert_data(data=data_dict, schema="Meal")

meal_to_recipe = pd.read_json(MEALTORECIPE)
meal_to_recipe.head(3)

filtered_df = meal_to_recipe[meal_to_recipe['meal_id'].isin(df_cleaned['id'])]
recipe_to_meal = filtered_df.rename(columns={"meal_id": "A", "recipe_id": "B"})
recipe_to_meal.head()

data_rm = recipe_to_meal.copy()
data_dict = data_rm.to_dict(orient='records')
# insert_data(data=data_dict, schema="_MealToRecipe")

CATEGORY = "../../app/data/processed/recipes_AI.json"
cat = pd.read_json(CATEGORY)
recipe_filtered = cat[cat['recipe_id'].isin(data_rm['B'])]
recipe_filtered
