==> app/notebooks/notebook_04_merge_recommender.ipynb <==
import pandas as pd
import numpy as np
import os
from typing import DefaultDict

PATH = "../data/processed/final_csvs/"

def load_data():
    dataset_dictionary = DefaultDict(pd.DataFrame)
    datasets = ["Meal.csv", "Category.csv", "Cuisines.csv", "Ingredient.csv", "CategoryToMeal.csv", "Recipe.csv", "CuisineToMeal.csv", "RecipeIngredient.csv"]
    for data in datasets:
        datafile = pd.read_csv(f'{PATH}{data}')
        name = data.split(".")[0].lower()
        dataset_dictionary[name] = datafile
    return dataset_dictionary


datasets = load_data()

meal = datasets["meal"][["meal_id", "name", "description"]]
category = datasets["category"]
category_to_meal = datasets["categorytomeal"]
cuisine = datasets["cuisines"]
ingredient = datasets["ingredient"]
cusines_to_meal = datasets["cuisinetomeal"]
recipeingredient = datasets["recipeingredient"]
recipe = datasets["recipe"][["recipe_id", "meal_id"]]

meal_df = meal.merge(cusines_to_meal, on="meal_id") \
    .merge(category_to_meal, on="meal_id") \
        .merge(recipe, on="meal_id") \
            .merge(recipeingredient, on="recipe_id") \
                .merge(ingredient, on="ingredient_id") \
                    .merge(cuisine, on="cuisines_id") \
                        .merge(category, on="category_id")

print(meal_df.columns)

from IPython.display import display

meal_df = meal_df[["meal_id", "name", "description", "category", "cuisines", "main_ingredient"]]
print(display(meal_df.head(4)))

meals_data = meal_df.copy()
meals_data = meals_data.fillna("")

# Group and aggregate
grouped_meals_data = meals_data.groupby("meal_id").agg({
    "name": "first",  # Take the first name for each meal
    "description": "first",
    "category": lambda x: list(set(x)),  # Get unique categories as a list
    "cuisines": lambda x: list(set(x)),  # Get unique cuisines as a list
    "main_ingredient": lambda x: list(set(x))  # Get unique ingredients as a list
}).reset_index()

grouped_meals_data = grouped_meals_data.drop_duplicates(subset=["meal_id"])
len(grouped_meals_data)

display(grouped_meals_data.head(4))

grouped_meals_data["cuisines"] = grouped_meals_data["cuisines"].apply(lambda x: " ".join(x))
grouped_meals_data["category"] = grouped_meals_data["category"].apply(lambda x: " ".join(x))
grouped_meals_data["main_ingredient"] = grouped_meals_data["main_ingredient"].apply(lambda x: " ".join(x))

display(grouped_meals_data.head(4))

data = grouped_meals_data.copy()

data["features"] = data["name"] +  " " + data["description"] + " " + data["category"] + " " + data["cuisines"] + " " + data["main_ingredient"]
data = data[["meal_id", "features"]]
data.head(4)

data.to_csv("../data/processed/final_csvs/meal_features.csv", index=False)

# ! pip install nltk tqdm datasets


import nltk
from tqdm import tqdm
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
from nltk.corpus import wordnet
from datasets import Dataset
import re

# nltk.download('stopwords')
# nltk.download('punkt_tab')
# nltk.download('wordnet')


data = pd.read_csv("../data/processed/final_csvs/meal_features.csv")
dataset = Dataset.from_pandas(data)
print(dataset[0])

stop_words = set(stopwords.words('english'))

def clean_text(dataset):
    text = dataset["features"]
    text = text.lower()
    text = word_tokenize(text)
    text = [word for word in text if word.isalnum()]
    text = [word for word in text if re.match(r'^[a-zA-Z0-9]+$', word)]
    text = [word for word in text if word not in stop_words]
    text = [WordNetLemmatizer().lemmatize(word) for word in text]
    dataset["features"] = " ".join(text)
    return dataset

dataset = dataset.map(clean_text, batched=False)
print(dataset)


dataset.save_to_disk("../data/processed/dataset")

# ! pip install spacy gensim
# ! python3 -m spacy download en_core_web_sm

sentence_array = dataset["features"]
# sentence_array = sentence_array[:, np.newaxis]
# sentence_array = [sent.split() for sentence in sentence_array for sent in sentence]

## Implementing tf-idf vectorizer

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel, cosine_similarity

vectorizer = TfidfVectorizer(ngram_range=(1, 3), min_df=1, analyzer='word', stop_words='english')
meal_tfidf_matrix = vectorizer.fit_transform(raw_documents=sentence_array)
matrix_LK = linear_kernel(meal_tfidf_matrix, meal_tfidf_matrix)

# Prepare user data

indexes = pd.Series(data["meal_id"])

user_dict = {
    "category": ["breakfast"],
    "cuisines": ["nigerian"],
    "intollerance": ["milk"],
    "allergies": ["pepper"],
    "preferences": ["vegetarian"]
}

user_data = pd.DataFrame([user_dict])

cols = user_data.columns

for col in cols:
    if isinstance(user_data[col][0], list):
        user_data[col] = user_data[col].apply(lambda x: " ".join(x))
    else:
        user_data[col] = user_data[col]

user_data["features"] = user_data["category"] + " " + user_data["cuisines"] + " " + user_data["preferences"]
user_data = user_data[["features"]]
user_data = Dataset.from_pandas(user_data)
user_data = user_data.map(clean_text, batched=False)
print(user_data)


user_content = user_data["features"]
user_tfidf_matrix = vectorizer.transform(user_content)
user_LK = cosine_similarity(user_tfidf_matrix, meal_tfidf_matrix).flatten()

user_dict

user_intolerances = set(user_dict["intollerance"][0].split())
user_allergies = set(user_dict["allergies"][0].split())

def contains_restricted(ingredients, restricted_set):
    """Check if any restricted item is present in the ingredients."""
    ingredients_set = set(ingredients.split())  # Assuming ingredients are space-separated
    return bool(ingredients_set & restricted_set)

filtered_meals = grouped_meals_data[~grouped_meals_data["main_ingredient"].apply(contains_restricted, args=(user_intolerances | user_allergies,))]
meal_rec = pd.DataFrame({"meal_id": indexes, "cosine_similarity": user_LK})
meal_rec = meal_rec.sort_values(by="cosine_similarity", ascending=False)

meal_rec = meal_rec.merge(filtered_meals, on="meal_id")
meal_rec
